Relocate merge utility and add test
