+++ src/options.js
@@ -1,4 +1,5 @@
 import Handlebars from 'handlebars';
+import { merge } from './utils';
 
 const defaults = {
   templates: {
@@ -9,24 +10,6 @@ const defaults = {
     partials  : ['src/partials/**/*']
   }
 };
-
-/**
- * Perform a deep merge of two objects.
- * @param {Object} target
- * @param {Object} source
- * @return {Object}
- * @example merge(defaults, options);
- */
-function merge (target, source) {
-  Object.keys(source).forEach(key => {
-    if (Object.isExtensible(source[key])) {
-      merge(target[key], source[key]);
-    } else {
-      Object.assign(target, { [key]: source[key] });
-    }
-  });
-  return target;
-}
 
 /**
  * Merge defaults into options.
